+++ app/src/main/java/com/sedsoftware/yaptalker/features/activetopics/ActiveTopicsPresenter.kt
@@ -67,8 +67,8 @@ class ActiveTopicsPresenter : BasePresenter<ActiveTopicsView>() {
 
   }
 
-  fun navigateToChosenTopic(pair: Pair<Int, Int>) {
-    router.navigateTo(NavigationScreens.CHOSEN_TOPIC_SCREEN, pair)
+  fun navigateToChosenTopic(triple: Triple<Int, Int, Int>) {
+    router.navigateTo(NavigationScreens.CHOSEN_TOPIC_SCREEN, triple)
   }
 
   fun goToFirstPage() {